+++ BetterFishing/Utilities/Items.cs
@@ -5,7 +5,6 @@ namespace BetterFishing
 {
     public class Items
     {
-        public static GameObject Hammer { get; internal set; }
         public static GameObject EmptyCrate { get; internal set; }
         public static GameObject SealingNails { get; internal set; }
         public static GameObject SpoonLure { get; internal set; }
@@ -14,22 +13,6 @@ namespace BetterFishing
         public static GameObject CrateSpoonLures { get; internal set; }
         public static GameObject CrateSwimbaitLures { get; internal set; }
         public static GameObject CrateTopwaterLures { get; internal set; }
-
-        internal static void InitializeHammer()
-        {
-            var itemHammer = Hammer.AddComponent<ShipItemHammer>();
-            itemHammer.holdDistance = 0.875f;
-            itemHammer.furniturePlaceHeight = 0.5f;
-            itemHammer.heldRotationOffset = -50;
-            itemHammer.mass = 1;
-            itemHammer.value = 120;
-            itemHammer.name = "hammer";
-            itemHammer.category = TransactionCategory.toolsAndSupplies;
-            itemHammer.inventoryScale = 1;
-            itemHammer.inventoryRotation = 90;
-            itemHammer.inventoryRotationX = -90;
-            itemHammer.floaterHeight = 1.6f;
-        }
 
         internal static void InitializeNails()
         {
@@ -59,7 +42,6 @@ namespace BetterFishing
 
         internal static void Initialize()
         {
-            InitializeHammer();
             InitializeNails();
             InitializeLure(SpoonLure, "good for catching tuna");
             InitializeLure(SwimbaitLure, "good for catching eel");